Set `want_handshake` on _data_ packet sent
WireGuard has a passive keep-alive mechanism
(https://www.wireguard.com/papers/wireguard.pdf#page=14). This one
mandates that - when you send a receive a data packet but no data packet
is sent as an organic response - a keep-alive packet is sent.

In addition, when sending a _data_ packet to the remote and we don't
receive a response within a certain time period, we initiate a new
handshake.

At present, `boringtun` sets the `want_handshake` property on every
outgoing packet, not just data packets. As a result, a keepalive packet
being sent to the remote as a result of not sending an organic response
would trigger the timer for expecting _yet another_ response from the
remote. The remote however won't send a keep-alive as a response to
receiving a keep-alive as that would lead to an endless loop of sending
keep-alives to each other.

By setting the `want_handshake` property on TimeLast_Data_PacketSent, we
preempt this.