fix: make `Timestamper` pure w.r.t. start time (#126)
The initial handshake packet of DH is vulnerable to replay attacks
because we do not yet have any knowledge about the client that is
attempting to connect. Such replay attacks could be used for DoS of
existing sessions. To protect against that, WireGuard uses TAI64N
timestamps in handshake packets. These timestamps are computed as an
offset of the current system time since the UNIX epoch. If a handshake
timestamp is equal or earlier to the last successful handshake, it is
discarded as a replay. The reasoning is explained in detail in
https://www.wireguard.com/protocol/#dos-mitigation.

Right now, these timestamps are computed using a UNIX timestamp offset
plus an `Instant` offset from when `Tunn` instance got created. This
allows for incremental updates of this timestamp where only a reference
to `Instant` is available.

When a peer discards its `Tunn` instance and then recreates it with the
exact same private key, the current impurity of `boringtun` causes us to
read a new UNIX timestamp. Given the strict ordering requirements
implemented in the handshake module, carefully controlling these
timestamps is critical.

To achieve that, we remove the call to `SystemTime::now` and require a
unix timestamp reference in the constructor of `Tunn` in addition to an
`Instant` reference, when that UNIX timestamp was taken. This allows us
to consistently compute the timestamp required for the handshake packets
without any additional syscalls within `boringtun`.

This should fix the `timestamp of packet is in the past; possible a
replay?` errors we have observed occasionally.